fix mobile Safari viewport height on full-screen dialog
