+++ src/lib/components/Dialog.svelte
@@ -17,6 +17,7 @@ Modal dialog component. Dispatches `open` and `close` events when state changes
 	import fsm from 'svelte-fsm';
 	import { createEventDispatcher, onMount } from 'svelte';
 	import { Icon } from '$lib/components';
+	import { setViewportHeight } from '$lib/actions/viewport';
 	import { toggleBodyScroll } from '$lib/helpers/scroll';
 
 	export let title = '';
@@ -52,7 +53,13 @@ Modal dialog component. Dispatches `open` and `close` events when state changes
 	$: toggleBodyScroll(open);
 </script>
 
-<dialog bind:this={dialog} class={fullScreen ? 'full-screen' : 'compact'} on:close={state.escaped} data-css-props>
+<dialog
+	bind:this={dialog}
+	class={fullScreen ? 'full-screen' : 'compact'}
+	use:setViewportHeight
+	on:close={state.escaped}
+	data-css-props
+>
 	<slot name="header">
 		<header>
 			<h5><slot name="title">{title}</slot></h5>
@@ -124,7 +131,7 @@ Modal dialog component. Dispatches `open` and `close` events when state changes
 	}
 
 	.full-screen {
-		height: 100vh;
+		height: var(--viewport-height, 100vh);
 		max-height: unset;
 
 		:global(:is(header, footer)) {
